refactor(citation-validator): extract AnchorMatcher and move to src/core/ (closes #28, closes #33)
- Extract AnchorMatcher from CitationValidator (companion to PathResolver in 14f4094)
- Move CitationValidator to src/core/CitationValidator/ subfolder alongside PathResolver and AnchorMatcher
- Add getLinks() to ParsedDocumentLike interface to fix TS error in CitationValidator.validateFile()
- Expose calculateRelativePath() and generatePathConversionSuggestion() as public delegates on CitationValidator
- Fix discriminated union narrowing in wiki-pre-resolved integration test (v.status !== 'error' guard)
- Remove unused imports (join from citation-validator-cache.test.js, vi from PathResolver.test.ts)
- Update factory + jact.ts imports to new core path
- Update all test imports to new core path
