+++ test/integration/citation-validator-cache.test.js
@@ -1,8 +1,8 @@
 import { existsSync, readFileSync } from "node:fs";
-import { dirname, join, resolve } from "node:path";
+import { dirname, resolve } from "node:path";
 import { fileURLToPath } from "node:url";
 import { beforeEach, describe, expect, it, vi } from "vitest";
-import { CitationValidator } from "../../src/CitationValidator.js";
+import { CitationValidator } from "../../src/core/CitationValidator/CitationValidator.js";
 import { MarkdownParser } from "../../src/core/MarkdownParser/index.js";
 import { ParsedFileCache } from "../../src/ParsedFileCache.js";
