+++ test/integration/citation-validator-single-citation.test.js
@@ -2,10 +2,10 @@ import { readFileSync } from "node:fs";
 import { dirname, resolve } from "node:path";
 import { fileURLToPath } from "node:url";
 import { beforeEach, describe, expect, it } from "vitest";
-import { CitationValidator } from "../../src/CitationValidator.js";
+import { CitationValidator } from "../../src/core/CitationValidator/CitationValidator.js";
 import { MarkdownParser } from "../../src/core/MarkdownParser/index.js";
-import { ParsedFileCache } from "../../src/ParsedFileCache.js";
 import { LinkObjectFactory } from "../../src/factories/LinkObjectFactory.js";
+import { ParsedFileCache } from "../../src/ParsedFileCache.js";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
